requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
edge-tts>=6.1.0
moviepy>=1.0.3
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
numpy>=1.24.0
mutagen>=1.47.0
